agreuge el modelo de usuario
